Move train_fail and deploy_fail to tasks.py to avoid circular imports later (they will soon be modified to call tasks).
